Ensure provider/ tests run before examples/ tests
Makes sure that provider/ tests are run in the test matrix, per language.

Brings .ci-mgmt.yaml and workflows in sync with `make ci-mgmt`.

Adopts test baselines re-recorded on CI.
