+++ .ci-mgmt.yaml
@@ -1,6 +1,6 @@
 provider: gcp
 gcp: true
-major-version: 6
+major-version: 7
 lint: false
 parallel: 2
 providerVersion: "github.com/hashicorp/terraform-provider-google-beta/version.ProviderVersion"
@@ -14,6 +14,7 @@ env:
   GOOGLE_REGION: "us-central1"
   GOOGLE_ZONE: "us-central1-a"
   PULUMI_MISSING_DOCS_ERROR: true
+  VERSION_PREFIX: 7.0.0
 upstream-provider-repo: terraform-provider-google-beta
 makeTemplate: bridged
 plugins:
@@ -34,3 +35,13 @@ javaGenVersion: v0.9.7
 runner:
   publish: pulumi-ubuntu-8core
   buildSdk: pulumi-ubuntu-8core
+actions:
+  preTest:
+
+    - name: make upstream
+      run: |
+        make upstream
+
+    - name: Run provider tests
+      run: |
+        cd provider && go test -v -json -count=1 -cover -timeout 2h -tags=${{ matrix.language }} -parallel 4 . 2>&1 | tee /tmp/gotest.log | gotestfmt
