+++ .github/workflows/pull-request.yml
@@ -34,6 +34,7 @@ env:
   SLACK_WEBHOOK_URL: ${{ secrets.SLACK_WEBHOOK_URL }}
   TF_APPEND_USER_AGENT: pulumi
   TRAVIS_OS_NAME: linux
+  VERSION_PREFIX: 7.0.0
 jobs:
   comment-on-pr:
     if: github.event.pull_request.head.repo.full_name != github.repository
