+++ .github/workflows/resync-build.yml
@@ -36,6 +36,7 @@ env:
   SLACK_WEBHOOK_URL: ${{ secrets.SLACK_WEBHOOK_URL }}
   TF_APPEND_USER_AGENT: pulumi
   TRAVIS_OS_NAME: linux
+  VERSION_PREFIX: 7.0.0
 jobs:
   resync_build:
     name: resync-build
