+++ provider/provider_test.go
@@ -5,7 +5,6 @@ package gcp
 import (
 	"context"
 	"fmt"
-	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -16,85 +15,6 @@ import (
 	pfbridge "github.com/pulumi/pulumi-terraform-bridge/pf/tfbridge"
 	pulumirpc "github.com/pulumi/pulumi/sdk/v3/proto/go"
 )
-
-func TestDNSRecordSet(t *testing.T) {
-	runTest(t, test(t, "test-programs/dns-recordset"))
-}
-
-func TestPubSubSubscription(t *testing.T) {
-	runTest(t, test(t, "test-programs/pubsub-subscription"))
-}
-
-func TestPubSubTopic(t *testing.T) {
-	runTest(t, test(t, "test-programs/pubsub-topic"))
-}
-
-func TestServiceAccount(t *testing.T) {
-	runTest(t, test(t, "test-programs/serviceaccount-account",
-		providertest.WithSkippedUpgradeTestMode(providertest.UpgradeTestMode_PreviewOnly,
-			"TODO[pulumi/providertest#7] PreviewOnly is confused about stack names"),
-	))
-}
-
-func TestStorageBucket(t *testing.T) {
-	runTest(t, test(t, "test-programs/storage-bucket"))
-}
-
-func TestStorageBucketObject(t *testing.T) {
-	t.Skipf("TODO[pulumi/providertest#2] skipping because Assets are not working yet")
-	runTest(t, test(t, "test-programs/storage-bucketobject"))
-}
-
-func TestSecretManagerSecret(t *testing.T) {
-	runTest(t, test(t, "test-programs/secretmanager-secret"))
-}
-
-func TestSqlUser(t *testing.T) {
-	runTest(t, test(t, "test-programs/sql-user"))
-}
-
-func TestBigQueryTable(t *testing.T) {
-	runTest(t, test(t, "test-programs/bigquery-table",
-		providertest.WithSkippedUpgradeTestMode(providertest.UpgradeTestMode_PreviewOnly,
-			"TODO[pulumi/providertest#7] PreviewOnly is confused about stack names"),
-	))
-}
-
-func TestComputeFirewall(t *testing.T) {
-	runTest(t, test(t, "test-programs/compute-firewall"))
-}
-
-func TestCloudFunction(t *testing.T) {
-	t.Skipf("TODO[pulumi/providertest#2] skipping because Assets are not working yet")
-	runTest(t, test(t, "test-programs/cloudfunctions-function"))
-}
-
-// Test programs that were automatically extracted from examples without autocorrection.
-func TestAutoExtractedPrograms(t *testing.T) {
-	type testCase struct {
-		program string
-	}
-
-	testCases := []testCase{
-		{"monitoring-alertpolicy-1"},
-		{"bigquery-datasetaccess-3"},
-		{"bigquery-routine-1"},
-		{"bigquery-routine-2"},
-		{"cloudrunv2-job-6"},
-		{"cloudrunv2-service-4"},
-		{"cloudrunv2-service-5"},
-		{"cloudrunv2-service-7"},
-		{"compute-disk-1"},
-		{"compute-disk-3"},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.program, func(t *testing.T) {
-			runTest(t, test(t, filepath.Join("test-programs", tc.program)))
-		})
-	}
-}
 
 func TestUpgradeCoverage(t *testing.T) {
 	providertest.ReportUpgradeCoverage(t)
@@ -111,8 +31,6 @@ func test(t *testing.T, dir string, opts ...providertest.Option) *providertest.P
 
 // This funcion inlines bits of ProviderTest.Run to avoid running E2E tests and instead focuses on
 // upgrade tests. Currently E2E tests are not able to run in this provider.
-//
-// TODO[pulumi/ci-mgmt#675] expose credentials so this just becomes pt.Run(t).
 func runTest(t *testing.T, pt *providertest.ProviderTest) {
 	t.Run("upgrade-snapshot", func(t *testing.T) {
 		t.Helper()
