+++ provider/testdata/recorded/TestProviderUpgrade/monitoring-alertpolicy-1/6.67.0/state.json
@@ -2,48 +2,52 @@
     "version": 3,
     "deployment": {
         "manifest": {
-            "time": "2023-10-10T10:41:36.494421-04:00",
-            "magic": "cdb46f20d3397474c4b7038ce63a3c40fb0da4b95458ef275b6ce689aec88d17",
-            "version": "v3.86.0"
+            "time": "2023-11-01T16:17:55.925752463Z",
+            "magic": "acdeccf9e5003385dafbac764321868e2deb68a98202938780c079c3d99ec55a",
+            "version": "v3.91.1"
         },
         "secrets_providers": {
             "type": "passphrase",
             "state": {
-                "salt": "v1:SsBk+X46P3A=:v1:VY5ndIczidA/KM/w:HW/eTpfDORarN+52SXTaf6HAiFhLKw=="
+                "salt": "v1:RRK+Dc5MxpQ=:v1:ZZaxZqYI6x67i5CL:5lQUQS5b4lUiuscWoyTVkT4aLClZWA=="
             }
         },
         "resources": [
             {
-                "urn": "urn:pulumi:p-it-antons-mac-monitoring-7b7cf1a6::test::pulumi:pulumi:Stack::test-p-it-antons-mac-monitoring-7b7cf1a6",
+                "urn": "urn:pulumi:p-it-runner-monitoring-44205fdd::test::pulumi:pulumi:Stack::test-p-it-runner-monitoring-44205fdd",
                 "custom": false,
                 "type": "pulumi:pulumi:Stack",
-                "created": "2023-10-10T14:41:31.949932Z",
-                "modified": "2023-10-10T14:41:31.949932Z"
+                "created": "2023-11-01T16:17:49.8241322Z",
+                "modified": "2023-11-01T16:17:49.8241322Z"
             },
             {
-                "urn": "urn:pulumi:p-it-antons-mac-monitoring-7b7cf1a6::test::pulumi:providers:gcp::provider",
+                "urn": "urn:pulumi:p-it-runner-monitoring-44205fdd::test::pulumi:providers:gcp::provider",
                 "custom": true,
-                "id": "f7c7ed9a-8cdb-4173-b83a-53b90f4a705e",
+                "id": "a4033452-2722-4008-8ee8-7b6b7450892b",
                 "type": "pulumi:providers:gcp",
                 "inputs": {
-                    "project": "pulumi-development",
-                    "version": "6.67.0"
+                    "project": "pulumi-ci-gcp-provider",
+                    "region": "us-central1",
+                    "version": "6.67.0",
+                    "zone": "us-central1-a"
                 },
                 "outputs": {
-                    "project": "pulumi-development",
-                    "version": "6.67.0"
+                    "project": "pulumi-ci-gcp-provider",
+                    "region": "us-central1",
+                    "version": "6.67.0",
+                    "zone": "us-central1-a"
                 },
-                "parent": "urn:pulumi:p-it-antons-mac-monitoring-7b7cf1a6::test::pulumi:pulumi:Stack::test-p-it-antons-mac-monitoring-7b7cf1a6",
+                "parent": "urn:pulumi:p-it-runner-monitoring-44205fdd::test::pulumi:pulumi:Stack::test-p-it-runner-monitoring-44205fdd",
                 "propertyDependencies": {
                     "project": null
                 },
-                "created": "2023-10-10T14:41:32.726597Z",
-                "modified": "2023-10-10T14:41:32.726597Z"
+                "created": "2023-11-01T16:17:51.89609098Z",
+                "modified": "2023-11-01T16:17:51.89609098Z"
             },
             {
-                "urn": "urn:pulumi:p-it-antons-mac-monitoring-7b7cf1a6::test::gcp:monitoring/alertPolicy:AlertPolicy::tested-resource",
+                "urn": "urn:pulumi:p-it-runner-monitoring-44205fdd::test::gcp:monitoring/alertPolicy:AlertPolicy::tested-resource",
                 "custom": true,
-                "id": "projects/pulumi-development/alertPolicies/3265642716557756309",
+                "id": "projects/pulumi-ci-gcp-provider/alertPolicies/15076563816038255889",
                 "type": "gcp:monitoring/alertPolicy:AlertPolicy",
                 "inputs": {
                     "__defaults": [
@@ -106,36 +110,36 @@
                                 "trigger": null
                             },
                             "displayName": "test condition",
-                            "name": "projects/pulumi-development/alertPolicies/3265642716557756309/conditions/3265642716557754896"
+                            "name": "projects/pulumi-ci-gcp-provider/alertPolicies/15076563816038255889/conditions/15076563816038254928"
                         }
                     ],
                     "creationRecords": [
                         {
-                            "mutateTime": "2023-10-10T14:41:33.273757984Z",
-                            "mutatedBy": "anton@pulumi.com"
+                            "mutateTime": "2023-11-01T16:17:52.537591232Z",
+                            "mutatedBy": "pulumi-ci@pulumi-ci-gcp-provider.iam.gserviceaccount.com"
                         }
                     ],
                     "displayName": "My Alert Policy",
                     "documentation": null,
                     "enabled": true,
-                    "id": "projects/pulumi-development/alertPolicies/3265642716557756309",
-                    "name": "projects/pulumi-development/alertPolicies/3265642716557756309",
+                    "id": "projects/pulumi-ci-gcp-provider/alertPolicies/15076563816038255889",
+                    "name": "projects/pulumi-ci-gcp-provider/alertPolicies/15076563816038255889",
                     "notificationChannels": [],
-                    "project": "pulumi-development",
+                    "project": "pulumi-ci-gcp-provider",
                     "userLabels": {
                         "foo": "bar"
                     }
                 },
-                "parent": "urn:pulumi:p-it-antons-mac-monitoring-7b7cf1a6::test::pulumi:pulumi:Stack::test-p-it-antons-mac-monitoring-7b7cf1a6",
-                "provider": "urn:pulumi:p-it-antons-mac-monitoring-7b7cf1a6::test::pulumi:providers:gcp::provider::f7c7ed9a-8cdb-4173-b83a-53b90f4a705e",
+                "parent": "urn:pulumi:p-it-runner-monitoring-44205fdd::test::pulumi:pulumi:Stack::test-p-it-runner-monitoring-44205fdd",
+                "provider": "urn:pulumi:p-it-runner-monitoring-44205fdd::test::pulumi:providers:gcp::provider::a4033452-2722-4008-8ee8-7b6b7450892b",
                 "propertyDependencies": {
                     "combiner": null,
                     "conditions": null,
                     "displayName": null,
                     "userLabels": null
                 },
-                "created": "2023-10-10T14:41:36.494349Z",
-                "modified": "2023-10-10T14:41:36.494349Z"
+                "created": "2023-11-01T16:17:55.925500321Z",
+                "modified": "2023-11-01T16:17:55.925500321Z"
             }
         ]
     }
