+++ provider/testdata/recorded/TestProviderUpgrade/serviceaccount-account/6.67.0/state.json
@@ -2,86 +2,92 @@
     "version": 3,
     "deployment": {
         "manifest": {
-            "time": "2023-10-06T15:34:25.649586-04:00",
-            "magic": "cdb46f20d3397474c4b7038ce63a3c40fb0da4b95458ef275b6ce689aec88d17",
-            "version": "v3.86.0"
+            "time": "2023-11-01T16:01:25.80892883Z",
+            "magic": "acdeccf9e5003385dafbac764321868e2deb68a98202938780c079c3d99ec55a",
+            "version": "v3.91.1"
         },
         "secrets_providers": {
             "type": "passphrase",
             "state": {
-                "salt": "v1:cQetygp1M/8=:v1:vuYXKvQFrWGMnCRK:N7Dr69X2ZOFY399c7LChZ1jeYZQfvA=="
+                "salt": "v1:U7Hs4IqI7e0=:v1:bil+vBlU9QGis4ts:g+A0Bn/RwfTamW0kPdm8FSto/3ksXg=="
             }
         },
         "resources": [
             {
-                "urn": "urn:pulumi:p-it-antons-mac-serviceacc-b3d4f1a7::test::pulumi:pulumi:Stack::test-p-it-antons-mac-serviceacc-b3d4f1a7",
+                "urn": "urn:pulumi:p-it-runner-serviceacc-17f48302::test::pulumi:pulumi:Stack::test-p-it-runner-serviceacc-17f48302",
                 "custom": false,
                 "type": "pulumi:pulumi:Stack",
-                "created": "2023-10-06T19:34:23.879811Z",
-                "modified": "2023-10-06T19:34:23.879811Z"
+                "created": "2023-11-01T16:01:22.27327712Z",
+                "modified": "2023-11-01T16:01:22.27327712Z"
             },
             {
-                "urn": "urn:pulumi:p-it-antons-mac-serviceacc-b3d4f1a7::test::pulumi:providers:gcp::provider",
+                "urn": "urn:pulumi:p-it-runner-serviceacc-17f48302::test::pulumi:providers:gcp::provider",
                 "custom": true,
-                "id": "09509483-a74f-4fc9-a37f-37065cffdfd5",
+                "id": "66d9ebda-9961-463a-a50e-3ea24bdf341c",
                 "type": "pulumi:providers:gcp",
                 "inputs": {
-                    "version": "6.67.0"
+                    "project": "pulumi-ci-gcp-provider",
+                    "region": "us-central1",
+                    "version": "6.67.0",
+                    "zone": "us-central1-a"
                 },
                 "outputs": {
-                    "version": "6.67.0"
+                    "project": "pulumi-ci-gcp-provider",
+                    "region": "us-central1",
+                    "version": "6.67.0",
+                    "zone": "us-central1-a"
                 },
-                "parent": "urn:pulumi:p-it-antons-mac-serviceacc-b3d4f1a7::test::pulumi:pulumi:Stack::test-p-it-antons-mac-serviceacc-b3d4f1a7",
+                "parent": "urn:pulumi:p-it-runner-serviceacc-17f48302::test::pulumi:pulumi:Stack::test-p-it-runner-serviceacc-17f48302",
                 "propertyDependencies": {
                     "version": []
                 },
-                "created": "2023-10-06T19:34:24.330369Z",
-                "modified": "2023-10-06T19:34:24.330369Z"
+                "created": "2023-11-01T16:01:24.244271311Z",
+                "modified": "2023-11-01T16:01:24.244271311Z"
             },
             {
-                "urn": "urn:pulumi:p-it-antons-mac-serviceacc-b3d4f1a7::test::pulumi:providers:str::default",
+                "urn": "urn:pulumi:p-it-runner-serviceacc-17f48302::test::pulumi:providers:str::default",
                 "custom": true,
-                "id": "98546a32-e737-4e0f-bdc0-7da737d809bc",
+                "id": "b329dc4c-b873-4379-bbc0-861b3fa2d54e",
                 "type": "pulumi:providers:str",
-                "created": "2023-10-06T19:34:24.36027Z",
-                "modified": "2023-10-06T19:34:24.36027Z"
+                "created": "2023-11-01T16:01:24.260991278Z",
+                "modified": "2023-11-01T16:01:24.260991278Z"
             },
             {
-                "urn": "urn:pulumi:p-it-antons-mac-serviceacc-b3d4f1a7::test::gcp:serviceAccount/account:Account::tested-resource",
+                "urn": "urn:pulumi:p-it-runner-serviceacc-17f48302::test::gcp:serviceAccount/account:Account::tested-resource",
                 "custom": true,
-                "id": "projects/pulumi-development/serviceAccounts/antons-mac-serviceacc-b3d4f1a7@pulumi-development.iam.gserviceaccount.com",
+                "id": "projects/pulumi-ci-gcp-provider/serviceAccounts/runner-serviceacc-17f48302@pulumi-ci-gcp-provider.iam.gserviceaccount.com",
                 "type": "gcp:serviceAccount/account:Account",
                 "inputs": {
                     "__defaults": [
                         "disabled"
                     ],
-                    "accountId": "antons-mac-serviceacc-b3d4f1a7",
+                    "accountId": "runner-serviceacc-17f48302",
                     "disabled": false,
                     "displayName": "Test Service Account",
-                    "project": "pulumi-development"
+                    "project": "pulumi-ci-gcp-provider"
                 },
                 "outputs": {
                     "__meta": "{\"e2bfb730-ecaa-11e6-8f88-34363bc7c4c0\":{\"create\":300000000000}}",
-                    "accountId": "antons-mac-serviceacc-b3d4f1a7",
+                    "accountId": "runner-serviceacc-17f48302",
                     "description": "",
                     "disabled": false,
                     "displayName": "Test Service Account",
-                    "email": "antons-mac-serviceacc-b3d4f1a7@pulumi-development.iam.gserviceaccount.com",
-                    "id": "projects/pulumi-development/serviceAccounts/antons-mac-serviceacc-b3d4f1a7@pulumi-development.iam.gserviceaccount.com",
-                    "member": "serviceAccount:antons-mac-serviceacc-b3d4f1a7@pulumi-development.iam.gserviceaccount.com",
-                    "name": "projects/pulumi-development/serviceAccounts/antons-mac-serviceacc-b3d4f1a7@pulumi-development.iam.gserviceaccount.com",
-                    "project": "pulumi-development",
-                    "uniqueId": "109782189447858634007"
+                    "email": "runner-serviceacc-17f48302@pulumi-ci-gcp-provider.iam.gserviceaccount.com",
+                    "id": "projects/pulumi-ci-gcp-provider/serviceAccounts/runner-serviceacc-17f48302@pulumi-ci-gcp-provider.iam.gserviceaccount.com",
+                    "member": "serviceAccount:runner-serviceacc-17f48302@pulumi-ci-gcp-provider.iam.gserviceaccount.com",
+                    "name": "projects/pulumi-ci-gcp-provider/serviceAccounts/runner-serviceacc-17f48302@pulumi-ci-gcp-provider.iam.gserviceaccount.com",
+                    "project": "pulumi-ci-gcp-provider",
+                    "uniqueId": "112825229045027648647"
                 },
-                "parent": "urn:pulumi:p-it-antons-mac-serviceacc-b3d4f1a7::test::pulumi:pulumi:Stack::test-p-it-antons-mac-serviceacc-b3d4f1a7",
-                "provider": "urn:pulumi:p-it-antons-mac-serviceacc-b3d4f1a7::test::pulumi:providers:gcp::provider::09509483-a74f-4fc9-a37f-37065cffdfd5",
+                "parent": "urn:pulumi:p-it-runner-serviceacc-17f48302::test::pulumi:pulumi:Stack::test-p-it-runner-serviceacc-17f48302",
+                "provider": "urn:pulumi:p-it-runner-serviceacc-17f48302::test::pulumi:providers:gcp::provider::66d9ebda-9961-463a-a50e-3ea24bdf341c",
                 "propertyDependencies": {
                     "accountId": null,
                     "displayName": null,
                     "project": null
                 },
-                "created": "2023-10-06T19:34:25.64955Z",
-                "modified": "2023-10-06T19:34:25.64955Z"
+                "created": "2023-11-01T16:01:25.80885367Z",
+                "modified": "2023-11-01T16:01:25.80885367Z"
             }
         ]
     }
